bug: 19221 - Do not hold whole message in memory
Missed another file in change 71019.  Brain not
functioning...

http://bugzilla.zimbra.com/show_bug.cgi?id=19221

+++ ZimbraOffline/src/java/com/zimbra/cs/mailbox/InitialSync.java
@@ -853,7 +853,7 @@ public class InitialSync {
         String digest = null;
         int size;
         try {
-            pm = new ParsedMessage(content, received, false);
+            pm = new ParsedMessage(content, (long) received, false);
             digest = pm.getRawDigest();
             size = pm.getRawSize();
         } catch (MessagingException e) {
